fix: split files exceeding 300 lines for downstream compatibility (#57)
Split 5 oversized files into smaller modules (all ≤250 lines) while
preserving all existing import paths via barrel re-exports.

Files split:
- src/lib/i18n/types.ts (726→7 files): split by domain (shell-theme,
  data, forms-nav, agentic-network, ops-strategy, umbrella + barrel)
- src/app/(dashboard)/dashboard/page.tsx (750→4 files): extract demo
  data, brain builders, widgets into sibling modules
- src/lib/component-catalog-data.ts (456→4 files): split entries into
  3 chunks (a/b/c) by category, barrel merges and re-exports
- src/app/(dashboard)/showcase/showcase-data.ts (345→3 files): split
  into viz+network and agentic data, barrel re-exports
- scripts/generate-component-docs.ts (1137→5 files): extract curated
  examples into 3 chunk files and helpers into separate module

Zero breaking changes — all original import paths continue to work.
Verified: typecheck ✓ lint ✓ all 22 files ≤250 lines ✓
