+++ src/lib/component-catalog-data.ts
@@ -1,4 +1,4 @@
-// Maranello component catalog data — 102 entries (single source of truth)
+// Maranello component catalog data — barrel re-export (split for maintainability)
 
 export interface CatalogEntry {
   name: string;
@@ -11,446 +11,8 @@ export interface CatalogEntry {
   propsInterface: string;
 }
 
-function c(name: string, slug: string, cat: string, desc: string, when: string, kw: string[]): CatalogEntry {
-  return { name, slug, category: cat, description: desc, whenToUse: when,
-    filePath: `${cat}/${slug}.tsx`, propsInterface: `${name}Props`, keywords: kw };
-}
+import { CATALOG_A } from './component-catalog-entries-a';
+import { CATALOG_B } from './component-catalog-entries-b';
+import { CATALOG_C } from './component-catalog-entries-c';
 
-export const CATALOG: CatalogEntry[] = [
-  // ── agentic (9) ──
-  c("MnActiveMissions", "mn-active-missions", "agentic",
-    "Dashboard panel tracking active AI agent missions and their real-time progress",
-    "Use to monitor running AI tasks, agent workflows, or background missions",
-    ["missions", "agent", "tasks", "missioni", "agente", "compiti", "active", "attive", "progress"]),
-  c("MnAgentTrace", "mn-agent-trace", "agentic",
-    "Trace viewer showing step-by-step execution of AI agent reasoning and actions",
-    "Use to debug or inspect how an AI agent arrived at its response",
-    ["trace", "agent", "debug", "traccia", "agente", "steps", "reasoning", "passaggi"]),
-  c("MnApprovalChain", "mn-approval-chain", "agentic",
-    "Sequential approval workflow showing sign-off status across stakeholders",
-    "Use for multi-step approval processes requiring ordered sign-offs",
-    ["approval", "chain", "workflow", "approvazione", "catena", "flusso", "sign-off"]),
-  c("MnProcessTimeline", "mn-process-timeline", "agentic",
-    "Horizontal or vertical multi-step workflow visualization with actor avatars and status indicators",
-    "Use to show end-to-end business processes with multiple actors handling different steps",
-    ["timeline", "process", "workflow", "steps", "actors", "processo", "flusso", "passaggi", "attori", "horizontal", "vertical"]),
-  c("MnAugmentedBrain", "mn-augmented-brain", "agentic",
-    "Visual representation of AI-augmented decision-making combining human and machine intelligence",
-    "Use to display the interplay between human insight and AI suggestions",
-    ["brain", "augmented", "AI", "cervello", "aumentato", "intelligence", "intelligenza"]),
-  c("MnChat", "mn-chat", "agentic",
-    "Real-time chat interface with message threading and AI assistant support",
-    "Use for conversational UIs, support chat, or AI copilot interactions",
-    ["chat", "message", "conversation", "messaggio", "conversazione", "assistente", "thread"]),
-  c("MnHubSpoke", "mn-hub-spoke", "agentic",
-    "Hub-and-spoke diagram visualizing central node relationships to satellite nodes",
-    "Use to show a central entity connected to multiple peripheral entities",
-    ["hub", "spoke", "diagram", "diagramma", "mozzo", "raggi", "network", "rete"]),
-  c("MnNeuralNodes", "mn-neural-nodes", "agentic",
-    "Interactive neural network node graph for visualizing model architectures",
-    "Use to render neural-network-style node graphs or AI pipeline topologies",
-    ["neural", "nodes", "network", "neurale", "nodi", "rete", "graph", "grafo", "AI"]),
-  c("MnBrain3D", "mn-brain-3d", "agentic",
-    "3D force-directed graph showing AI agents as glowing spheres with animated edges",
-    "Use to visualize agent networks, delegation flows, and inter-agent relationships in 3D",
-    ["brain", "3d", "graph", "cervello", "grafo", "agents", "agenti", "force", "network", "rete"]),
-  c("MnAugmentedBrainV2", "mn-augmented-brain-v2", "agentic",
-    "2D canvas brain visualization with hub-satellite layout, synapse particles, and stats bar",
-    "Use for an augmented brain panel showing a central hub connected to task and agent nodes",
-    ["brain", "augmented", "cervello", "aumentato", "hub", "synapse", "sinapsi", "canvas"]),
-
-  // ── data-display (12) ──
-  c("MnAvatar", "mn-avatar", "data-display",
-    "User avatar with initials fallback, status indicator, and size variants",
-    "Use to represent users or entities with photo, initials, or icon",
-    ["avatar", "user", "profile", "utente", "profilo", "image", "immagine", "photo"]),
-  c("MnBadge", "mn-badge", "data-display",
-    "Status indicator badge with CVA variants for severity levels",
-    "ALWAYS use for status indicators, counts, or labels. Do NOT create custom colored spans or inline status elements.",
-    ["badge", "status", "tag", "etichetta", "stato", "label", "count", "conteggio"]),
-  c("MnDataTable", "mn-data-table", "data-display",
-    "Full-featured data table with sorting, filtering, pagination, and row selection",
-    "ALWAYS use for ANY tabular data display. Do NOT create custom <table> elements or grid layouts. Provides sorting, filtering, pagination, and row selection out of the box.",
-    ["table", "data", "tabella", "dati", "grid", "griglia", "sort", "filter", "pagination"]),
-  c("MnDetailPanel", "mn-detail-panel", "data-display",
-    "Slide-out panel showing detailed information for a selected record",
-    "ALWAYS use for slide-out detail views. Do NOT use Sheet, Dialog, or custom sidebars. Provides consistent header, close button, scroll handling, and theme integration.",
-    ["detail", "panel", "dettaglio", "pannello", "drawer", "sidebar", "info", "informazioni"]),
-  c("MnFlipCounter", "mn-flip-counter", "data-display",
-    "Animated flip-card counter for displaying changing numeric values",
-    "Use for animated number transitions in dashboards or scoreboards",
-    ["counter", "flip", "animation", "contatore", "animazione", "number", "numero", "ticker"]),
-  c("MnIcon", "mn-icon", "data-display",
-    "SVG icon renderer with consistent sizing and theme-aware coloring",
-    "Use to display icons from the Maranello icon set",
-    ["icon", "svg", "icona", "symbol", "simbolo", "glyph"]),
-  c("MnKpiScorecard", "mn-kpi-scorecard", "data-display",
-    "KPI scorecard displaying key metrics with trend indicators and sparklines",
-    "Use to highlight important business metrics with delta and trend context",
-    ["kpi", "scorecard", "metric", "metrica", "indicatore", "trend", "performance", "prestazione"]),
-  c("MnProgressRing", "mn-progress-ring", "data-display",
-    "Circular progress indicator with percentage label and animated fill",
-    "Use to show completion percentage of a task or process",
-    ["progress", "ring", "circular", "progresso", "anello", "circolare", "percentage"]),
-  c("MnSourceCards", "mn-source-cards", "data-display",
-    "Card grid displaying data source metadata with connection status",
-    "Use to present data sources, integrations, or API connections as cards",
-    ["source", "card", "fonte", "scheda", "integration", "integrazione", "connection", "connessione"]),
-  c("MnSpinner", "mn-spinner", "data-display",
-    "Loading spinner with size and color variants for async state indication",
-    "Use to indicate loading or processing state in any context",
-    ["spinner", "loading", "caricamento", "loader", "wait", "attesa", "progress"]),
-  c("MnTokenMeter", "mn-token-meter", "data-display",
-    "Token usage meter showing consumption against quota with color thresholds",
-    "Use to display AI token consumption, API quotas, or usage limits",
-    ["token", "meter", "usage", "consumo", "quota", "limit", "limite", "misuratore"]),
-  c("MnUserTable", "mn-user-table", "data-display",
-    "User management table with role badges, status filters, and bulk actions",
-    "Use to list and manage users with role-based access information",
-    ["user", "table", "utente", "tabella", "role", "ruolo", "management", "gestione"]),
-
-  // ── data-viz (14) ──
-  c("MnBudgetTreemap", "mn-budget-treemap", "data-viz",
-    "Treemap chart visualizing budget allocation across hierarchical categories",
-    "Use to show proportional budget distribution with drill-down capability",
-    ["treemap", "budget", "bilancio", "chart", "grafico", "allocation", "allocazione"]),
-  c("MnBulletChart", "mn-bullet-chart", "data-viz",
-    "Bullet chart comparing a measured value against target ranges",
-    "Use to display performance metrics against qualitative thresholds",
-    ["bullet", "chart", "grafico", "target", "obiettivo", "performance", "range", "intervallo"]),
-  c("MnChart", "mn-chart", "data-viz",
-    "General-purpose chart wrapper supporting line, bar, area, and pie chart types",
-    "Use as the base chart component for standard data visualizations",
-    ["chart", "grafico", "line", "bar", "area", "pie", "torta", "linea", "barra"]),
-  c("MnCohortGrid", "mn-cohort-grid", "data-viz",
-    "Cohort analysis grid showing retention or behavior across time cohorts",
-    "Use to visualize user cohort retention or temporal behavior patterns",
-    ["cohort", "grid", "coorte", "griglia", "retention", "ritenzione", "analysis", "analisi"]),
-  c("MnConfidenceChart", "mn-confidence-chart", "data-viz",
-    "Chart displaying confidence intervals and probability distributions",
-    "Use to visualize prediction confidence, error ranges, or probability bands",
-    ["confidence", "chart", "grafico", "confidenza", "interval", "intervallo", "probability"]),
-  c("MnCostTimeline", "mn-cost-timeline", "data-viz",
-    "Timeline chart showing cost evolution with cumulative trend line",
-    "Use to track expenses or costs over time with running totals",
-    ["cost", "timeline", "costo", "linea temporale", "expense", "spesa", "trend", "cumulative"]),
-  c("MnFunnel", "mn-funnel", "data-viz",
-    "Funnel chart showing progressive stage-to-stage conversion rates",
-    "Use to visualize sales, conversion, or pipeline funnels",
-    ["funnel", "imbuto", "conversion", "conversione", "pipeline", "stages", "fasi", "vendite"]),
-  c("MnGauge", "mn-gauge", "data-viz",
-    "Full circular gauge dial with needle indicator and configurable ranges",
-    "Use to display a single metric as a dial gauge with color-coded zones",
-    ["gauge", "indicatore", "dial", "quadrante", "meter", "misuratore", "needle", "lancetta"]),
-  c("MnHalfGauge", "mn-half-gauge", "data-viz",
-    "Semi-circular gauge for compact metric display with threshold zones",
-    "Use when a full gauge is too large and a half-circle fits the layout",
-    ["gauge", "half", "semi", "indicatore", "mezzo", "semicircolare", "compact"]),
-  c("MnHbar", "mn-hbar", "data-viz",
-    "Horizontal bar chart with grouped or stacked bar support",
-    "Use for horizontal bar charts and compact pipeline/funnel views. Preferred over MnFunnel for compact layouts. Do NOT create custom div-based bar charts.",
-    ["bar", "horizontal", "chart", "barra", "orizzontale", "grafico", "comparison", "confronto"]),
-  c("MnHeatmap", "mn-heatmap", "data-viz",
-    "Matrix heatmap with color-coded cells and interactive tooltips",
-    "Use to reveal patterns across two dimensions via color intensity",
-    ["heatmap", "mappa di calore", "matrix", "matrice", "color", "colore", "intensity"]),
-  c("MnPipelineRanking", "mn-pipeline-ranking", "data-viz",
-    "Ranked pipeline chart showing deal or item scores in order",
-    "Use to rank and compare pipeline items by a scoring metric",
-    ["pipeline", "ranking", "classifica", "rank", "score", "punteggio", "deal", "ordine"]),
-  c("MnSpeedometer", "mn-speedometer", "data-viz",
-    "Speedometer-style gauge with animated needle and RPM-inspired design",
-    "Use for a high-impact single-value gauge with automotive aesthetics",
-    ["speedometer", "tachimetro", "gauge", "speed", "velocità", "needle", "lancetta"]),
-  c("MnWaterfall", "mn-waterfall", "data-viz",
-    "Waterfall chart showing cumulative positive and negative value changes",
-    "Use to break down how incremental values contribute to a total",
-    ["waterfall", "cascata", "chart", "grafico", "cumulative", "cumulativo", "breakdown"]),
-
-  // ── feedback (6) ──
-  c("MnActivityFeed", "mn-activity-feed", "feedback",
-    "Chronological activity feed with grouped events and actor avatars",
-    "Use to display a timeline of user actions, system events, or audit entries",
-    ["activity", "feed", "attività", "flusso", "timeline", "cronologia", "events", "eventi"]),
-  c("MnModal", "mn-modal", "feedback",
-    "Accessible modal dialog with backdrop, focus trap, and size variants",
-    "Use for confirmations, forms, or content that requires focused attention",
-    ["modal", "dialog", "modale", "dialogo", "popup", "overlay", "confirm", "conferma"]),
-  c("MnNotificationCenter", "mn-notification-center", "feedback",
-    "Notification center panel with categorized alerts and read/unread state",
-    "Use to aggregate and manage app-wide notifications in one place",
-    ["notification", "center", "notifica", "centro", "alert", "avviso", "bell", "campana"]),
-  c("MnStateScaffold", "mn-state-scaffold", "feedback",
-    "Empty-state scaffold rendering contextual illustrations and call-to-action",
-    "Use for empty, error, or loading states to guide users toward next steps",
-    ["state", "scaffold", "stato", "placeholder", "empty", "vuoto", "error", "errore"]),
-  c("MnStreamingText", "mn-streaming-text", "feedback",
-    "Streaming text display with typewriter animation for AI-generated content",
-    "Use when rendering text that arrives token-by-token from an LLM",
-    ["streaming", "text", "testo", "typewriter", "AI", "LLM", "animation", "animazione"]),
-  c("MnToast", "mn-toast", "feedback",
-    "Toast notification with auto-dismiss, stacking, and severity variants",
-    "Use for brief, non-blocking feedback after user actions",
-    ["toast", "notification", "notifica", "snackbar", "alert", "avviso", "message", "messaggio"]),
-
-  // ── financial (2) ──
-  c("MnAgentCostBreakdown", "mn-agent-cost-breakdown", "financial",
-    "Breakdown panel detailing AI agent costs by model, tokens, and task",
-    "Use to itemize and attribute costs of AI agent operations",
-    ["cost", "breakdown", "costo", "scomposizione", "agent", "agente", "billing", "fatturazione"]),
-  c("MnFinOps", "mn-finops", "financial",
-    "Financial operations dashboard with spend tracking and budget alerts",
-    "Use to monitor cloud and AI spend with budget guardrails",
-    ["finops", "financial", "finanziario", "operations", "operazioni", "spend", "spesa", "budget"]),
-
-  // ── forms (11) ──
-  c("MnAsyncSelect", "mn-async-select", "forms",
-    "Async select dropdown with server-side search and debounced loading",
-    "Use for selects that load options from an API as the user types",
-    ["select", "async", "dropdown", "selezione", "asincrono", "search", "ricerca", "autocomplete"]),
-  c("MnCalendarRange", "mn-calendar-range", "forms",
-    "Calendar-based range selector with preset period shortcuts",
-    "Use to pick contiguous date ranges from a visual calendar",
-    ["calendar", "range", "calendario", "intervallo", "date", "data", "period", "periodo"]),
-  c("MnDatePicker", "mn-date-picker", "forms",
-    "Single-date picker with calendar popup and manual input support",
-    "Use to select a single date from a calendar or by typing",
-    ["date", "picker", "data", "selettore", "calendar", "calendario", "input"]),
-  c("MnDateRangePicker", "mn-date-range-picker", "forms",
-    "Dual-calendar range picker with comparison period support",
-    "Use to select start and end dates for reporting or filtering",
-    ["date", "range", "data", "intervallo", "picker", "selettore", "from", "to", "da", "a"]),
-  c("MnFilterPanel", "mn-filter-panel", "forms",
-    "Collapsible filter panel with multi-facet controls and apply/reset actions",
-    "Use to provide advanced multi-criteria filtering for data views",
-    ["filter", "panel", "filtro", "pannello", "facet", "criteria", "criteri", "search", "ricerca"]),
-  c("MnFormField", "mn-form-field", "forms",
-    "Form field wrapper providing label, validation, helper text, and error state",
-    "Use to wrap any input with consistent label, hint, and error display",
-    ["form", "field", "modulo", "campo", "input", "label", "etichetta", "validation", "validazione"]),
-  c("MnLogin", "mn-login", "forms",
-    "Login form with email/password fields, SSO options, and forgot-password link",
-    "Use as the primary authentication entry point",
-    ["login", "auth", "accesso", "autenticazione", "password", "sign-in", "email"]),
-  c("MnProfile", "mn-profile", "forms",
-    "User profile form with avatar upload, personal info, and preference fields",
-    "Use for user settings pages where profile data can be viewed and edited",
-    ["profile", "user", "profilo", "utente", "settings", "impostazioni", "account"]),
-  c("MnSearchDrawer", "mn-search-drawer", "forms",
-    "Full-height search drawer with real-time results and category filters",
-    "Use for global search experiences that overlay the main content",
-    ["search", "drawer", "ricerca", "cassetto", "find", "trova", "query", "results", "risultati"]),
-  c("MnToggleSwitch", "mn-toggle-switch", "forms",
-    "Toggle switch with on/off labels and accessible keyboard interaction",
-    "Use for binary settings that take effect immediately",
-    ["toggle", "switch", "interruttore", "on", "off", "boolean", "setting", "impostazione"]),
-  c("MnVoiceInput", "mn-voice-input", "forms",
-    "Voice input button with speech-to-text transcription and waveform feedback",
-    "Use to add voice-based input alongside or instead of typing",
-    ["voice", "input", "voce", "speech", "parlato", "microphone", "microfono", "STT"]),
-
-  // ── layout (8) ──
-  c("MnAdminShell", "mn-admin-shell", "layout",
-    "Admin layout shell with sidebar navigation, header, and content area",
-    "Use as the top-level layout for admin or back-office applications",
-    ["admin", "shell", "layout", "amministrazione", "guscio", "sidebar", "barra laterale"]),
-  c("MnDashboardRenderer", "mn-dashboard-renderer", "layout",
-    "Dynamic dashboard renderer that lays out widget configs into a grid",
-    "Use to render user-configured dashboards from a JSON widget definition",
-    ["dashboard", "renderer", "cruscotto", "rendering", "widget", "dynamic", "dinamico"]),
-  c("MnDashboardStrip", "mn-dashboard-strip", "layout",
-    "Horizontal strip of compact dashboard KPI cards",
-    "ALWAYS use for dashboard KPI strips and top-of-page summary metrics. Do NOT create custom metric card divs with hardcoded colors.",
-    ["dashboard", "strip", "cruscotto", "striscia", "kpi", "summary", "riepilogo"]),
-  c("MnDashboard", "mn-dashboard", "layout",
-    "Pre-composed dashboard layout with configurable widget slots",
-    "Use to build standard dashboards with a header, sidebar, and grid area",
-    ["dashboard", "cruscotto", "layout", "grid", "griglia", "widget", "panel", "pannello"]),
-  c("MnGridLayout", "mn-grid-layout", "layout",
-    "Responsive grid layout manager with drag-and-drop widget repositioning",
-    "Use when users need to rearrange widgets within a grid",
-    ["grid", "layout", "griglia", "disposizione", "drag", "drop", "responsive", "reattivo"]),
-  c("MnHeaderShell", "mn-header-shell", "layout",
-    "App header bar with logo slot, navigation, search, and user menu",
-    "Use as the top-level header for all application pages",
-    ["header", "shell", "intestazione", "guscio", "navbar", "barra di navigazione", "top"]),
-  c("MnSectionCard", "mn-section-card", "layout",
-    "Card container with title bar, actions slot, and collapsible body",
-    "Use to group related content into a visually distinct section",
-    ["section", "card", "sezione", "scheda", "container", "contenitore", "panel", "pannello"]),
-  c("MnSettingsPanel", "mn-settings-panel", "layout",
-    "Settings form panel with categorized option groups and save/cancel controls",
-    "Use to present application or module settings in a structured panel",
-    ["settings", "panel", "impostazioni", "pannello", "config", "configurazione", "options"]),
-
-  // ── navigation (5) ──
-  c("MnBreadcrumb", "mn-breadcrumb", "navigation",
-    "Breadcrumb trail with separator customization and overflow truncation",
-    "Use to show hierarchical page location and enable upward navigation",
-    ["breadcrumb", "navigation", "briciole di pane", "navigazione", "path", "percorso"]),
-  c("MnCommandPalette", "mn-command-palette", "navigation",
-    "Keyboard-driven command palette for quick actions and navigation",
-    "Use to give power users a fast command and search interface",
-    ["command", "palette", "comando", "tavolozza", "search", "ricerca", "shortcut", "scorciatoia"]),
-  c("MnSectionNav", "mn-section-nav", "navigation",
-    "Vertical section navigation sidebar with nested collapsible groups",
-    "Use for intra-page or intra-module vertical navigation menus",
-    ["section", "nav", "sezione", "navigazione", "sidebar", "barra laterale", "menu"]),
-  c("MnStepper", "mn-stepper", "navigation",
-    "Multi-step wizard stepper with progress tracking and step validation",
-    "Use to guide users through a sequential multi-step workflow",
-    ["stepper", "wizard", "step", "passo", "procedura guidata", "progress", "workflow"]),
-  c("MnTabs", "mn-tabs", "navigation",
-    "Tabbed interface with accessible keyboard navigation and lazy panel loading",
-    "Use to organize content into mutually exclusive tabbed panels",
-    ["tabs", "tab", "schede", "scheda", "panel", "pannello", "switch", "navigazione"]),
-
-  // ── network (10) ──
-  c("MnDeploymentTable", "mn-deployment-table", "network",
-    "Deployment status table listing services, versions, and health checks",
-    "Use to monitor deployed services with their current version and status",
-    ["deployment", "table", "distribuzione", "tabella", "service", "servizio", "version"]),
-  c("MnMap", "mn-map", "network",
-    "Interactive geographic map with marker clusters and region overlays",
-    "Use to display location-based data on a geographic map",
-    ["map", "mappa", "geography", "geografia", "location", "posizione", "marker", "pin"]),
-  c("MnMeshNetworkCanvas", "mn-mesh-network-canvas", "network",
-    "Canvas-rendered mesh network visualization for large-scale topologies",
-    "Use for high-performance rendering of complex mesh network graphs",
-    ["mesh", "canvas", "rete", "tela", "network", "topology", "topologia", "graph", "grafo"]),
-  c("MnMeshNetworkCard", "mn-mesh-network-card", "network",
-    "Summary card for a single node within a mesh network view",
-    "Use to display metadata and status of an individual mesh network node",
-    ["mesh", "card", "rete", "scheda", "node", "nodo", "network", "status", "stato"]),
-  c("MnMeshNetworkToolbar", "mn-mesh-network-toolbar", "network",
-    "Toolbar for mesh network controls: zoom, filter, layout, and export",
-    "Use alongside MnMeshNetwork to provide user controls for the graph",
-    ["mesh", "toolbar", "rete", "barra strumenti", "controls", "controlli", "zoom", "filter"]),
-  c("MnMeshNetwork", "mn-mesh-network", "network",
-    "Composite mesh network view combining canvas, cards, and toolbar",
-    "Use to render a complete mesh network topology with interactive controls",
-    ["mesh", "network", "rete", "topology", "topologia", "graph", "grafo", "nodes", "nodi"]),
-  c("MnNetworkMessages", "mn-network-messages", "network",
-    "Message log panel for network communication events and packet traces",
-    "Use to display real-time or historical network message traffic",
-    ["network", "messages", "rete", "messaggi", "log", "packet", "pacchetto", "traffic"]),
-  c("MnOrgChart", "mn-org-chart", "network",
-    "Organizational chart with hierarchical node layout and expand/collapse",
-    "Use to visualize company or team hierarchies with reporting lines",
-    ["org", "chart", "organigramma", "grafico", "hierarchy", "gerarchia", "organization"]),
-  c("MnSocialGraph", "mn-social-graph", "network",
-    "Social network graph showing relationships and connection strength",
-    "Use to map social or professional relationships between entities",
-    ["social", "graph", "sociale", "grafo", "network", "rete", "relationship", "relazione"]),
-  c("MnSystemStatus", "mn-system-status", "network",
-    "System status dashboard showing service health with uptime indicators",
-    "Use to provide an at-a-glance operational health overview",
-    ["system", "status", "sistema", "stato", "health", "salute", "uptime", "operativo"]),
-
-  // ── ops (8) ──
-  c("MnAuditLog", "mn-audit-log", "ops",
-    "Paginated audit log table with actor, action, and timestamp columns",
-    "Use to display a chronological record of system or user actions",
-    ["audit", "log", "registro", "controllo", "history", "cronologia", "trace", "traccia"]),
-  c("MnBinnacle", "mn-binnacle", "ops",
-    "Operational binnacle log combining events, metrics, and alerts in one stream",
-    "Use to unify operational signals into a single chronological view",
-    ["binnacle", "log", "giornale di bordo", "registro", "operations", "operazioni", "events"]),
-  c("MnEntityWorkbench", "mn-entity-workbench", "ops",
-    "Workbench view for inspecting and editing a single business entity",
-    "Use to provide a full editing surface for a complex domain entity",
-    ["entity", "workbench", "entità", "banco di lavoro", "edit", "modifica", "detail"]),
-  c("MnFacetWorkbench", "mn-facet-workbench", "ops",
-    "Multi-facet workbench for exploring data across configurable dimensions",
-    "Use to let analysts slice data by multiple facets simultaneously",
-    ["facet", "workbench", "sfaccettatura", "banco di lavoro", "dimension", "dimensione"]),
-  c("MnGantt", "mn-gantt", "ops",
-    "Gantt chart with task dependencies, milestones, and drag-to-reschedule",
-    "Use to plan and track project schedules with task timelines",
-    ["gantt", "chart", "grafico", "project", "progetto", "timeline", "cronologia", "schedule"]),
-  c("MnInstrumentBinnacle", "mn-instrument-binnacle", "ops",
-    "Instrument cluster binnacle with gauges, indicators, and real-time telemetry",
-    "Use to present multiple live instruments in a dashboard cockpit view",
-    ["instrument", "binnacle", "strumento", "giornale di bordo", "cockpit", "cruscotto"]),
-  c("MnKanbanBoard", "mn-kanban-board", "ops",
-    "Kanban board with draggable cards, swimlanes, and WIP limits",
-    "Use to manage workflow stages with visual card-based task tracking",
-    ["kanban", "board", "lavagna", "task", "compito", "drag", "workflow", "flusso", "agile"]),
-  c("MnNightJobs", "mn-night-jobs", "ops",
-    "Night jobs monitor showing scheduled batch job runs and their outcomes",
-    "Use to track overnight or scheduled batch processing jobs",
-    ["night", "jobs", "lavori notturni", "batch", "lotto", "schedule", "programmazione", "cron"]),
-
-  // ── strategy (11) ──
-  c("MnBcgMatrix", "mn-bcg-matrix", "strategy",
-    "BCG growth-share matrix with quadrant labels and bubble plot",
-    "Use to classify business units or products by growth and market share",
-    ["bcg", "matrix", "matrice", "growth", "crescita", "share", "quota", "strategy", "strategia"]),
-  c("MnBusinessModelCanvas", "mn-business-model-canvas", "strategy",
-    "Business Model Canvas with nine editable building blocks",
-    "Use to map out a business model across value proposition, channels, and more",
-    ["business", "model", "canvas", "modello", "tela", "strategy", "strategia", "value"]),
-  c("MnCustomerJourneyMap", "mn-customer-journey-map", "strategy",
-    "Customer journey map with touchpoints, emotions, and channel lanes",
-    "Use to visualize end-to-end customer experience across channels",
-    ["customer", "journey", "map", "cliente", "viaggio", "mappa", "touchpoint", "esperienza"]),
-  c("MnCustomerJourney", "mn-customer-journey", "strategy",
-    "Simplified customer journey flow showing sequential interaction steps",
-    "Use for a compact, linear view of key customer interaction stages",
-    ["customer", "journey", "cliente", "viaggio", "flow", "flusso", "stages", "fasi"]),
-  c("MnDecisionMatrix", "mn-decision-matrix", "strategy",
-    "Weighted decision matrix for scoring options against criteria",
-    "Use to objectively compare alternatives with weighted scoring",
-    ["decision", "matrix", "decisione", "matrice", "criteria", "criteri", "score", "punteggio"]),
-  c("MnNineBoxMatrix", "mn-nine-box-matrix", "strategy",
-    "Nine-box talent matrix plotting performance vs. potential",
-    "Use for talent reviews and succession planning discussions",
-    ["nine-box", "matrix", "matrice", "talent", "talento", "performance", "potential", "HR"]),
-  c("MnOkr", "mn-okr", "strategy",
-    "OKR tracker with objectives, key results, and progress indicators",
-    "Use to set, track, and review Objectives and Key Results",
-    ["okr", "objectives", "obiettivi", "key results", "risultati chiave", "goals", "traguardi"]),
-  c("MnPorterFiveForces", "mn-porter-five-forces", "strategy",
-    "Porter's Five Forces diagram with editable force descriptions",
-    "Use to analyze competitive dynamics using Porter's framework",
-    ["porter", "five forces", "cinque forze", "strategy", "strategia", "competition"]),
-  c("MnRiskMatrix", "mn-risk-matrix", "strategy",
-    "Risk matrix heatmap plotting likelihood vs. impact with editable risks",
-    "Use to assess and prioritize risks by probability and severity",
-    ["risk", "matrix", "rischio", "matrice", "likelihood", "probabilità", "impact", "impatto"]),
-  c("MnStrategyCanvas", "mn-strategy-canvas", "strategy",
-    "Strategy canvas line chart comparing value curves across competitors",
-    "Use to contrast strategic profiles using Blue Ocean value curves",
-    ["strategy", "canvas", "strategia", "tela", "value curve", "curva di valore", "comparison"]),
-  c("MnSwot", "mn-swot", "strategy",
-    "SWOT analysis grid with four editable quadrants",
-    "Use to conduct a structured SWOT analysis with collaborative editing",
-    ["swot", "strengths", "weaknesses", "opportunities", "threats", "forze", "debolezze"]),
-
-  // ── theme (6) ──
-  c("MnA11yFab", "mn-a11y-fab", "theme",
-    "Floating action button toggling accessibility overlays and settings",
-    "Use to give users quick access to accessibility tools from any page",
-    ["a11y", "fab", "accessibility", "accessibilità", "floating", "button", "pulsante"]),
-  c("MnA11y", "mn-a11y", "theme",
-    "Accessibility settings panel with contrast, font-size, and motion controls",
-    "Use to let users customize accessibility preferences app-wide",
-    ["a11y", "accessibility", "accessibilità", "contrast", "contrasto", "font", "motion"]),
-  c("MnDropdownMenu", "mn-dropdown-menu", "theme",
-    "Dropdown menu with nested sub-menus, icons, and keyboard navigation",
-    "Use for context menus or action menus triggered by a button",
-    ["dropdown", "menu", "tendina", "menù", "context", "contesto", "action", "azione"]),
-  { name: "MnManettino", slug: "mn-manettino", category: "theme",
-    description: "Ferrari-inspired rotary mode selector with haptic-style detent positions",
-    whenToUse: "Use as a themed mode or level selector with rotary dial interaction",
-    filePath: "theme/mn-ferrari-control.tsx", propsInterface: "MnManettinoProps",
-    keywords: ["manettino", "rotary", "ferrari", "selector", "selettore", "dial", "ghiera"] },
-  c("MnThemeRotary", "mn-theme-rotary", "theme",
-    "Rotary dial for selecting the active color theme with preview swatches",
-    "Use to let users browse and switch themes via a physical-dial metaphor",
-    ["theme", "rotary", "tema", "rotante", "color", "colore", "palette", "tavolozza"]),
-  c("MnThemeToggle", "mn-theme-toggle", "theme",
-    "Light/dark theme toggle switch with system-preference detection",
-    "Use to switch between light, dark, and system color schemes",
-    ["theme", "toggle", "tema", "interruttore", "dark", "light", "scuro", "chiaro", "mode"]),
-];
+export const CATALOG: CatalogEntry[] = [...CATALOG_A, ...CATALOG_B, ...CATALOG_C];
