+++ src/lib/component-catalog-entries-a.ts
@@ -1,0 +1,158 @@
+import type { CatalogEntry } from './component-catalog-data';
+
+function c(name: string, slug: string, cat: string, desc: string, when: string, kw: string[]): CatalogEntry {
+  return { name, slug, category: cat, description: desc, whenToUse: when,
+    filePath: `${cat}/${slug}.tsx`, propsInterface: `${name}Props`, keywords: kw };
+}
+
+export const CATALOG_A: CatalogEntry[] = [
+  // ── agentic (10) ──
+  c("MnActiveMissions", "mn-active-missions", "agentic",
+    "Dashboard panel tracking active AI agent missions and their real-time progress",
+    "Use to monitor running AI tasks, agent workflows, or background missions",
+    ["missions", "agent", "tasks", "missioni", "agente", "compiti", "active", "attive", "progress"]),
+  c("MnAgentTrace", "mn-agent-trace", "agentic",
+    "Trace viewer showing step-by-step execution of AI agent reasoning and actions",
+    "Use to debug or inspect how an AI agent arrived at its response",
+    ["trace", "agent", "debug", "traccia", "agente", "steps", "reasoning", "passaggi"]),
+  c("MnApprovalChain", "mn-approval-chain", "agentic",
+    "Sequential approval workflow showing sign-off status across stakeholders",
+    "Use for multi-step approval processes requiring ordered sign-offs",
+    ["approval", "chain", "workflow", "approvazione", "catena", "flusso", "sign-off"]),
+  c("MnProcessTimeline", "mn-process-timeline", "agentic",
+    "Horizontal or vertical multi-step workflow visualization with actor avatars and status indicators",
+    "Use to show end-to-end business processes with multiple actors handling different steps",
+    ["timeline", "process", "workflow", "steps", "actors", "processo", "flusso", "passaggi", "attori", "horizontal", "vertical"]),
+  c("MnAugmentedBrain", "mn-augmented-brain", "agentic",
+    "Visual representation of AI-augmented decision-making combining human and machine intelligence",
+    "Use to display the interplay between human insight and AI suggestions",
+    ["brain", "augmented", "AI", "cervello", "aumentato", "intelligence", "intelligenza"]),
+  c("MnChat", "mn-chat", "agentic",
+    "Real-time chat interface with message threading and AI assistant support",
+    "Use for conversational UIs, support chat, or AI copilot interactions",
+    ["chat", "message", "conversation", "messaggio", "conversazione", "assistente", "thread"]),
+  c("MnHubSpoke", "mn-hub-spoke", "agentic",
+    "Hub-and-spoke diagram visualizing central node relationships to satellite nodes",
+    "Use to show a central entity connected to multiple peripheral entities",
+    ["hub", "spoke", "diagram", "diagramma", "mozzo", "raggi", "network", "rete"]),
+  c("MnNeuralNodes", "mn-neural-nodes", "agentic",
+    "Interactive neural network node graph for visualizing model architectures",
+    "Use to render neural-network-style node graphs or AI pipeline topologies",
+    ["neural", "nodes", "network", "neurale", "nodi", "rete", "graph", "grafo", "AI"]),
+  c("MnBrain3D", "mn-brain-3d", "agentic",
+    "3D force-directed graph showing AI agents as glowing spheres with animated edges",
+    "Use to visualize agent networks, delegation flows, and inter-agent relationships in 3D",
+    ["brain", "3d", "graph", "cervello", "grafo", "agents", "agenti", "force", "network", "rete"]),
+  c("MnAugmentedBrainV2", "mn-augmented-brain-v2", "agentic",
+    "2D canvas brain visualization with hub-satellite layout, synapse particles, and stats bar",
+    "Use for an augmented brain panel showing a central hub connected to task and agent nodes",
+    ["brain", "augmented", "cervello", "aumentato", "hub", "synapse", "sinapsi", "canvas"]),
+
+  // ── data-display (12) ──
+  c("MnAvatar", "mn-avatar", "data-display",
+    "User avatar with initials fallback, status indicator, and size variants",
+    "Use to represent users or entities with photo, initials, or icon",
+    ["avatar", "user", "profile", "utente", "profilo", "image", "immagine", "photo"]),
+  c("MnBadge", "mn-badge", "data-display",
+    "Status indicator badge with CVA variants for severity levels",
+    "ALWAYS use for status indicators, counts, or labels. Do NOT create custom colored spans or inline status elements.",
+    ["badge", "status", "tag", "etichetta", "stato", "label", "count", "conteggio"]),
+  c("MnDataTable", "mn-data-table", "data-display",
+    "Full-featured data table with sorting, filtering, pagination, and row selection",
+    "ALWAYS use for ANY tabular data display. Do NOT create custom <table> elements or grid layouts. Provides sorting, filtering, pagination, and row selection out of the box.",
+    ["table", "data", "tabella", "dati", "grid", "griglia", "sort", "filter", "pagination"]),
+  c("MnDetailPanel", "mn-detail-panel", "data-display",
+    "Slide-out panel showing detailed information for a selected record",
+    "ALWAYS use for slide-out detail views. Do NOT use Sheet, Dialog, or custom sidebars. Provides consistent header, close button, scroll handling, and theme integration.",
+    ["detail", "panel", "dettaglio", "pannello", "drawer", "sidebar", "info", "informazioni"]),
+  c("MnFlipCounter", "mn-flip-counter", "data-display",
+    "Animated flip-card counter for displaying changing numeric values",
+    "Use for animated number transitions in dashboards or scoreboards",
+    ["counter", "flip", "animation", "contatore", "animazione", "number", "numero", "ticker"]),
+  c("MnIcon", "mn-icon", "data-display",
+    "SVG icon renderer with consistent sizing and theme-aware coloring",
+    "Use to display icons from the Maranello icon set",
+    ["icon", "svg", "icona", "symbol", "simbolo", "glyph"]),
+  c("MnKpiScorecard", "mn-kpi-scorecard", "data-display",
+    "KPI scorecard displaying key metrics with trend indicators and sparklines",
+    "Use to highlight important business metrics with delta and trend context",
+    ["kpi", "scorecard", "metric", "metrica", "indicatore", "trend", "performance", "prestazione"]),
+  c("MnProgressRing", "mn-progress-ring", "data-display",
+    "Circular progress indicator with percentage label and animated fill",
+    "Use to show completion percentage of a task or process",
+    ["progress", "ring", "circular", "progresso", "anello", "circolare", "percentage"]),
+  c("MnSourceCards", "mn-source-cards", "data-display",
+    "Card grid displaying data source metadata with connection status",
+    "Use to present data sources, integrations, or API connections as cards",
+    ["source", "card", "fonte", "scheda", "integration", "integrazione", "connection", "connessione"]),
+  c("MnSpinner", "mn-spinner", "data-display",
+    "Loading spinner with size and color variants for async state indication",
+    "Use to indicate loading or processing state in any context",
+    ["spinner", "loading", "caricamento", "loader", "wait", "attesa", "progress"]),
+  c("MnTokenMeter", "mn-token-meter", "data-display",
+    "Token usage meter showing consumption against quota with color thresholds",
+    "Use to display AI token consumption, API quotas, or usage limits",
+    ["token", "meter", "usage", "consumo", "quota", "limit", "limite", "misuratore"]),
+  c("MnUserTable", "mn-user-table", "data-display",
+    "User management table with role badges, status filters, and bulk actions",
+    "Use to list and manage users with role-based access information",
+    ["user", "table", "utente", "tabella", "role", "ruolo", "management", "gestione"]),
+
+  // ── data-viz (14) ──
+  c("MnBudgetTreemap", "mn-budget-treemap", "data-viz",
+    "Treemap chart visualizing budget allocation across hierarchical categories",
+    "Use to show proportional budget distribution with drill-down capability",
+    ["treemap", "budget", "bilancio", "chart", "grafico", "allocation", "allocazione"]),
+  c("MnBulletChart", "mn-bullet-chart", "data-viz",
+    "Bullet chart comparing a measured value against target ranges",
+    "Use to display performance metrics against qualitative thresholds",
+    ["bullet", "chart", "grafico", "target", "obiettivo", "performance", "range", "intervallo"]),
+  c("MnChart", "mn-chart", "data-viz",
+    "General-purpose chart wrapper supporting line, bar, area, and pie chart types",
+    "Use as the base chart component for standard data visualizations",
+    ["chart", "grafico", "line", "bar", "area", "pie", "torta", "linea", "barra"]),
+  c("MnCohortGrid", "mn-cohort-grid", "data-viz",
+    "Cohort analysis grid showing retention or behavior across time cohorts",
+    "Use to visualize user cohort retention or temporal behavior patterns",
+    ["cohort", "grid", "coorte", "griglia", "retention", "ritenzione", "analysis", "analisi"]),
+  c("MnConfidenceChart", "mn-confidence-chart", "data-viz",
+    "Chart displaying confidence intervals and probability distributions",
+    "Use to visualize prediction confidence, error ranges, or probability bands",
+    ["confidence", "chart", "grafico", "confidenza", "interval", "intervallo", "probability"]),
+  c("MnCostTimeline", "mn-cost-timeline", "data-viz",
+    "Timeline chart showing cost evolution with cumulative trend line",
+    "Use to track expenses or costs over time with running totals",
+    ["cost", "timeline", "costo", "linea temporale", "expense", "spesa", "trend", "cumulative"]),
+  c("MnFunnel", "mn-funnel", "data-viz",
+    "Funnel chart showing progressive stage-to-stage conversion rates",
+    "Use to visualize sales, conversion, or pipeline funnels",
+    ["funnel", "imbuto", "conversion", "conversione", "pipeline", "stages", "fasi", "vendite"]),
+  c("MnGauge", "mn-gauge", "data-viz",
+    "Full circular gauge dial with needle indicator and configurable ranges",
+    "Use to display a single metric as a dial gauge with color-coded zones",
+    ["gauge", "indicatore", "dial", "quadrante", "meter", "misuratore", "needle", "lancetta"]),
+  c("MnHalfGauge", "mn-half-gauge", "data-viz",
+    "Semi-circular gauge for compact metric display with threshold zones",
+    "Use when a full gauge is too large and a half-circle fits the layout",
+    ["gauge", "half", "semi", "indicatore", "mezzo", "semicircolare", "compact"]),
+  c("MnHbar", "mn-hbar", "data-viz",
+    "Horizontal bar chart with grouped or stacked bar support",
+    "Use for horizontal bar charts and compact pipeline/funnel views. Preferred over MnFunnel for compact layouts. Do NOT create custom div-based bar charts.",
+    ["bar", "horizontal", "chart", "barra", "orizzontale", "grafico", "comparison", "confronto"]),
+  c("MnHeatmap", "mn-heatmap", "data-viz",
+    "Matrix heatmap with color-coded cells and interactive tooltips",
+    "Use to reveal patterns across two dimensions via color intensity",
+    ["heatmap", "mappa di calore", "matrix", "matrice", "color", "colore", "intensity"]),
+  c("MnPipelineRanking", "mn-pipeline-ranking", "data-viz",
+    "Ranked pipeline chart showing deal or item scores in order",
+    "Use to rank and compare pipeline items by a scoring metric",
+    ["pipeline", "ranking", "classifica", "rank", "score", "punteggio", "deal", "ordine"]),
+  c("MnSpeedometer", "mn-speedometer", "data-viz",
+    "Speedometer-style gauge with animated needle and RPM-inspired design",
+    "Use for a high-impact single-value gauge with automotive aesthetics",
+    ["speedometer", "tachimetro", "gauge", "speed", "velocità", "needle", "lancetta"]),
+  c("MnWaterfall", "mn-waterfall", "data-viz",
+    "Waterfall chart showing cumulative positive and negative value changes",
+    "Use to break down how incremental values contribute to a total",
+    ["waterfall", "cascata", "chart", "grafico", "cumulative", "cumulativo", "breakdown"]),
+];
