+++ src/lib/component-catalog-entries-b.ts
@@ -1,0 +1,188 @@
+import type { CatalogEntry } from './component-catalog-data';
+
+function c(name: string, slug: string, cat: string, desc: string, when: string, kw: string[]): CatalogEntry {
+  return { name, slug, category: cat, description: desc, whenToUse: when,
+    filePath: `${cat}/${slug}.tsx`, propsInterface: `${name}Props`, keywords: kw };
+}
+
+export const CATALOG_B: CatalogEntry[] = [
+  // ── feedback (6) ──
+  c("MnActivityFeed", "mn-activity-feed", "feedback",
+    "Chronological activity feed with grouped events and actor avatars",
+    "Use to display a timeline of user actions, system events, or audit entries",
+    ["activity", "feed", "attività", "flusso", "timeline", "cronologia", "events", "eventi"]),
+  c("MnModal", "mn-modal", "feedback",
+    "Accessible modal dialog with backdrop, focus trap, and size variants",
+    "Use for confirmations, forms, or content that requires focused attention",
+    ["modal", "dialog", "modale", "dialogo", "popup", "overlay", "confirm", "conferma"]),
+  c("MnNotificationCenter", "mn-notification-center", "feedback",
+    "Notification center panel with categorized alerts and read/unread state",
+    "Use to aggregate and manage app-wide notifications in one place",
+    ["notification", "center", "notifica", "centro", "alert", "avviso", "bell", "campana"]),
+  c("MnStateScaffold", "mn-state-scaffold", "feedback",
+    "Empty-state scaffold rendering contextual illustrations and call-to-action",
+    "Use for empty, error, or loading states to guide users toward next steps",
+    ["state", "scaffold", "stato", "placeholder", "empty", "vuoto", "error", "errore"]),
+  c("MnStreamingText", "mn-streaming-text", "feedback",
+    "Streaming text display with typewriter animation for AI-generated content",
+    "Use when rendering text that arrives token-by-token from an LLM",
+    ["streaming", "text", "testo", "typewriter", "AI", "LLM", "animation", "animazione"]),
+  c("MnToast", "mn-toast", "feedback",
+    "Toast notification with auto-dismiss, stacking, and severity variants",
+    "Use for brief, non-blocking feedback after user actions",
+    ["toast", "notification", "notifica", "snackbar", "alert", "avviso", "message", "messaggio"]),
+
+  // ── financial (2) ──
+  c("MnAgentCostBreakdown", "mn-agent-cost-breakdown", "financial",
+    "Breakdown panel detailing AI agent costs by model, tokens, and task",
+    "Use to itemize and attribute costs of AI agent operations",
+    ["cost", "breakdown", "costo", "scomposizione", "agent", "agente", "billing", "fatturazione"]),
+  c("MnFinOps", "mn-finops", "financial",
+    "Financial operations dashboard with spend tracking and budget alerts",
+    "Use to monitor cloud and AI spend with budget guardrails",
+    ["finops", "financial", "finanziario", "operations", "operazioni", "spend", "spesa", "budget"]),
+
+  // ── forms (11) ──
+  c("MnAsyncSelect", "mn-async-select", "forms",
+    "Async select dropdown with server-side search and debounced loading",
+    "Use for selects that load options from an API as the user types",
+    ["select", "async", "dropdown", "selezione", "asincrono", "search", "ricerca", "autocomplete"]),
+  c("MnCalendarRange", "mn-calendar-range", "forms",
+    "Calendar-based range selector with preset period shortcuts",
+    "Use to pick contiguous date ranges from a visual calendar",
+    ["calendar", "range", "calendario", "intervallo", "date", "data", "period", "periodo"]),
+  c("MnDatePicker", "mn-date-picker", "forms",
+    "Single-date picker with calendar popup and manual input support",
+    "Use to select a single date from a calendar or by typing",
+    ["date", "picker", "data", "selettore", "calendar", "calendario", "input"]),
+  c("MnDateRangePicker", "mn-date-range-picker", "forms",
+    "Dual-calendar range picker with comparison period support",
+    "Use to select start and end dates for reporting or filtering",
+    ["date", "range", "data", "intervallo", "picker", "selettore", "from", "to", "da", "a"]),
+  c("MnFilterPanel", "mn-filter-panel", "forms",
+    "Collapsible filter panel with multi-facet controls and apply/reset actions",
+    "Use to provide advanced multi-criteria filtering for data views",
+    ["filter", "panel", "filtro", "pannello", "facet", "criteria", "criteri", "search", "ricerca"]),
+  c("MnFormField", "mn-form-field", "forms",
+    "Form field wrapper providing label, validation, helper text, and error state",
+    "Use to wrap any input with consistent label, hint, and error display",
+    ["form", "field", "modulo", "campo", "input", "label", "etichetta", "validation", "validazione"]),
+  c("MnLogin", "mn-login", "forms",
+    "Login form with email/password fields, SSO options, and forgot-password link",
+    "Use as the primary authentication entry point",
+    ["login", "auth", "accesso", "autenticazione", "password", "sign-in", "email"]),
+  c("MnProfile", "mn-profile", "forms",
+    "User profile form with avatar upload, personal info, and preference fields",
+    "Use for user settings pages where profile data can be viewed and edited",
+    ["profile", "user", "profilo", "utente", "settings", "impostazioni", "account"]),
+  c("MnSearchDrawer", "mn-search-drawer", "forms",
+    "Full-height search drawer with real-time results and category filters",
+    "Use for global search experiences that overlay the main content",
+    ["search", "drawer", "ricerca", "cassetto", "find", "trova", "query", "results", "risultati"]),
+  c("MnToggleSwitch", "mn-toggle-switch", "forms",
+    "Toggle switch with on/off labels and accessible keyboard interaction",
+    "Use for binary settings that take effect immediately",
+    ["toggle", "switch", "interruttore", "on", "off", "boolean", "setting", "impostazione"]),
+  c("MnVoiceInput", "mn-voice-input", "forms",
+    "Voice input button with speech-to-text transcription and waveform feedback",
+    "Use to add voice-based input alongside or instead of typing",
+    ["voice", "input", "voce", "speech", "parlato", "microphone", "microfono", "STT"]),
+
+  // ── layout (8) ──
+  c("MnAdminShell", "mn-admin-shell", "layout",
+    "Admin layout shell with sidebar navigation, header, and content area",
+    "Use as the top-level layout for admin or back-office applications",
+    ["admin", "shell", "layout", "amministrazione", "guscio", "sidebar", "barra laterale"]),
+  c("MnDashboardRenderer", "mn-dashboard-renderer", "layout",
+    "Dynamic dashboard renderer that lays out widget configs into a grid",
+    "Use to render user-configured dashboards from a JSON widget definition",
+    ["dashboard", "renderer", "cruscotto", "rendering", "widget", "dynamic", "dinamico"]),
+  c("MnDashboardStrip", "mn-dashboard-strip", "layout",
+    "Horizontal strip of compact dashboard KPI cards",
+    "ALWAYS use for dashboard KPI strips and top-of-page summary metrics. Do NOT create custom metric card divs with hardcoded colors.",
+    ["dashboard", "strip", "cruscotto", "striscia", "kpi", "summary", "riepilogo"]),
+  c("MnDashboard", "mn-dashboard", "layout",
+    "Pre-composed dashboard layout with configurable widget slots",
+    "Use to build standard dashboards with a header, sidebar, and grid area",
+    ["dashboard", "cruscotto", "layout", "grid", "griglia", "widget", "panel", "pannello"]),
+  c("MnGridLayout", "mn-grid-layout", "layout",
+    "Responsive grid layout manager with drag-and-drop widget repositioning",
+    "Use when users need to rearrange widgets within a grid",
+    ["grid", "layout", "griglia", "disposizione", "drag", "drop", "responsive", "reattivo"]),
+  c("MnHeaderShell", "mn-header-shell", "layout",
+    "App header bar with logo slot, navigation, search, and user menu",
+    "Use as the top-level header for all application pages",
+    ["header", "shell", "intestazione", "guscio", "navbar", "barra di navigazione", "top"]),
+  c("MnSectionCard", "mn-section-card", "layout",
+    "Card container with title bar, actions slot, and collapsible body",
+    "Use to group related content into a visually distinct section",
+    ["section", "card", "sezione", "scheda", "container", "contenitore", "panel", "pannello"]),
+  c("MnSettingsPanel", "mn-settings-panel", "layout",
+    "Settings form panel with categorized option groups and save/cancel controls",
+    "Use to present application or module settings in a structured panel",
+    ["settings", "panel", "impostazioni", "pannello", "config", "configurazione", "options"]),
+
+  // ── navigation (5) ──
+  c("MnBreadcrumb", "mn-breadcrumb", "navigation",
+    "Breadcrumb trail with separator customization and overflow truncation",
+    "Use to show hierarchical page location and enable upward navigation",
+    ["breadcrumb", "navigation", "briciole di pane", "navigazione", "path", "percorso"]),
+  c("MnCommandPalette", "mn-command-palette", "navigation",
+    "Keyboard-driven command palette for quick actions and navigation",
+    "Use to give power users a fast command and search interface",
+    ["command", "palette", "comando", "tavolozza", "search", "ricerca", "shortcut", "scorciatoia"]),
+  c("MnSectionNav", "mn-section-nav", "navigation",
+    "Vertical section navigation sidebar with nested collapsible groups",
+    "Use for intra-page or intra-module vertical navigation menus",
+    ["section", "nav", "sezione", "navigazione", "sidebar", "barra laterale", "menu"]),
+  c("MnStepper", "mn-stepper", "navigation",
+    "Multi-step wizard stepper with progress tracking and step validation",
+    "Use to guide users through a sequential multi-step workflow",
+    ["stepper", "wizard", "step", "passo", "procedura guidata", "progress", "workflow"]),
+  c("MnTabs", "mn-tabs", "navigation",
+    "Tabbed interface with accessible keyboard navigation and lazy panel loading",
+    "Use to organize content into mutually exclusive tabbed panels",
+    ["tabs", "tab", "schede", "scheda", "panel", "pannello", "switch", "navigazione"]),
+
+  // ── network (10) ──
+  c("MnDeploymentTable", "mn-deployment-table", "network",
+    "Deployment status table listing services, versions, and health checks",
+    "Use to monitor deployed services with their current version and status",
+    ["deployment", "table", "distribuzione", "tabella", "service", "servizio", "version"]),
+  c("MnMap", "mn-map", "network",
+    "Interactive geographic map with marker clusters and region overlays",
+    "Use to display location-based data on a geographic map",
+    ["map", "mappa", "geography", "geografia", "location", "posizione", "marker", "pin"]),
+  c("MnMeshNetworkCanvas", "mn-mesh-network-canvas", "network",
+    "Canvas-rendered mesh network visualization for large-scale topologies",
+    "Use for high-performance rendering of complex mesh network graphs",
+    ["mesh", "canvas", "rete", "tela", "network", "topology", "topologia", "graph", "grafo"]),
+  c("MnMeshNetworkCard", "mn-mesh-network-card", "network",
+    "Summary card for a single node within a mesh network view",
+    "Use to display metadata and status of an individual mesh network node",
+    ["mesh", "card", "rete", "scheda", "node", "nodo", "network", "status", "stato"]),
+  c("MnMeshNetworkToolbar", "mn-mesh-network-toolbar", "network",
+    "Toolbar for mesh network controls: zoom, filter, layout, and export",
+    "Use alongside MnMeshNetwork to provide user controls for the graph",
+    ["mesh", "toolbar", "rete", "barra strumenti", "controls", "controlli", "zoom", "filter"]),
+  c("MnMeshNetwork", "mn-mesh-network", "network",
+    "Composite mesh network view combining canvas, cards, and toolbar",
+    "Use to render a complete mesh network topology with interactive controls",
+    ["mesh", "network", "rete", "topology", "topologia", "graph", "grafo", "nodes", "nodi"]),
+  c("MnNetworkMessages", "mn-network-messages", "network",
+    "Message log panel for network communication events and packet traces",
+    "Use to display real-time or historical network message traffic",
+    ["network", "messages", "rete", "messaggi", "log", "packet", "pacchetto", "traffic"]),
+  c("MnOrgChart", "mn-org-chart", "network",
+    "Organizational chart with hierarchical node layout and expand/collapse",
+    "Use to visualize company or team hierarchies with reporting lines",
+    ["org", "chart", "organigramma", "grafico", "hierarchy", "gerarchia", "organization"]),
+  c("MnSocialGraph", "mn-social-graph", "network",
+    "Social network graph showing relationships and connection strength",
+    "Use to map social or professional relationships between entities",
+    ["social", "graph", "sociale", "grafo", "network", "rete", "relationship", "relazione"]),
+  c("MnSystemStatus", "mn-system-status", "network",
+    "System status dashboard showing service health with uptime indicators",
+    "Use to provide an at-a-glance operational health overview",
+    ["system", "status", "sistema", "stato", "health", "salute", "uptime", "operativo"]),
+];
