+++ src/lib/component-catalog-entries-c.ts
@@ -1,0 +1,115 @@
+import type { CatalogEntry } from './component-catalog-data';
+
+function c(name: string, slug: string, cat: string, desc: string, when: string, kw: string[]): CatalogEntry {
+  return { name, slug, category: cat, description: desc, whenToUse: when,
+    filePath: `${cat}/${slug}.tsx`, propsInterface: `${name}Props`, keywords: kw };
+}
+
+export const CATALOG_C: CatalogEntry[] = [
+  // ── ops (8) ──
+  c("MnAuditLog", "mn-audit-log", "ops",
+    "Paginated audit log table with actor, action, and timestamp columns",
+    "Use to display a chronological record of system or user actions",
+    ["audit", "log", "registro", "controllo", "history", "cronologia", "trace", "traccia"]),
+  c("MnBinnacle", "mn-binnacle", "ops",
+    "Operational binnacle log combining events, metrics, and alerts in one stream",
+    "Use to unify operational signals into a single chronological view",
+    ["binnacle", "log", "giornale di bordo", "registro", "operations", "operazioni", "events"]),
+  c("MnEntityWorkbench", "mn-entity-workbench", "ops",
+    "Workbench view for inspecting and editing a single business entity",
+    "Use to provide a full editing surface for a complex domain entity",
+    ["entity", "workbench", "entità", "banco di lavoro", "edit", "modifica", "detail"]),
+  c("MnFacetWorkbench", "mn-facet-workbench", "ops",
+    "Multi-facet workbench for exploring data across configurable dimensions",
+    "Use to let analysts slice data by multiple facets simultaneously",
+    ["facet", "workbench", "sfaccettatura", "banco di lavoro", "dimension", "dimensione"]),
+  c("MnGantt", "mn-gantt", "ops",
+    "Gantt chart with task dependencies, milestones, and drag-to-reschedule",
+    "Use to plan and track project schedules with task timelines",
+    ["gantt", "chart", "grafico", "project", "progetto", "timeline", "cronologia", "schedule"]),
+  c("MnInstrumentBinnacle", "mn-instrument-binnacle", "ops",
+    "Instrument cluster binnacle with gauges, indicators, and real-time telemetry",
+    "Use to present multiple live instruments in a dashboard cockpit view",
+    ["instrument", "binnacle", "strumento", "giornale di bordo", "cockpit", "cruscotto"]),
+  c("MnKanbanBoard", "mn-kanban-board", "ops",
+    "Kanban board with draggable cards, swimlanes, and WIP limits",
+    "Use to manage workflow stages with visual card-based task tracking",
+    ["kanban", "board", "lavagna", "task", "compito", "drag", "workflow", "flusso", "agile"]),
+  c("MnNightJobs", "mn-night-jobs", "ops",
+    "Night jobs monitor showing scheduled batch job runs and their outcomes",
+    "Use to track overnight or scheduled batch processing jobs",
+    ["night", "jobs", "lavori notturni", "batch", "lotto", "schedule", "programmazione", "cron"]),
+
+  // ── strategy (11) ──
+  c("MnBcgMatrix", "mn-bcg-matrix", "strategy",
+    "BCG growth-share matrix with quadrant labels and bubble plot",
+    "Use to classify business units or products by growth and market share",
+    ["bcg", "matrix", "matrice", "growth", "crescita", "share", "quota", "strategy", "strategia"]),
+  c("MnBusinessModelCanvas", "mn-business-model-canvas", "strategy",
+    "Business Model Canvas with nine editable building blocks",
+    "Use to map out a business model across value proposition, channels, and more",
+    ["business", "model", "canvas", "modello", "tela", "strategy", "strategia", "value"]),
+  c("MnCustomerJourneyMap", "mn-customer-journey-map", "strategy",
+    "Customer journey map with touchpoints, emotions, and channel lanes",
+    "Use to visualize end-to-end customer experience across channels",
+    ["customer", "journey", "map", "cliente", "viaggio", "mappa", "touchpoint", "esperienza"]),
+  c("MnCustomerJourney", "mn-customer-journey", "strategy",
+    "Simplified customer journey flow showing sequential interaction steps",
+    "Use for a compact, linear view of key customer interaction stages",
+    ["customer", "journey", "cliente", "viaggio", "flow", "flusso", "stages", "fasi"]),
+  c("MnDecisionMatrix", "mn-decision-matrix", "strategy",
+    "Weighted decision matrix for scoring options against criteria",
+    "Use to objectively compare alternatives with weighted scoring",
+    ["decision", "matrix", "decisione", "matrice", "criteria", "criteri", "score", "punteggio"]),
+  c("MnNineBoxMatrix", "mn-nine-box-matrix", "strategy",
+    "Nine-box talent matrix plotting performance vs. potential",
+    "Use for talent reviews and succession planning discussions",
+    ["nine-box", "matrix", "matrice", "talent", "talento", "performance", "potential", "HR"]),
+  c("MnOkr", "mn-okr", "strategy",
+    "OKR tracker with objectives, key results, and progress indicators",
+    "Use to set, track, and review Objectives and Key Results",
+    ["okr", "objectives", "obiettivi", "key results", "risultati chiave", "goals", "traguardi"]),
+  c("MnPorterFiveForces", "mn-porter-five-forces", "strategy",
+    "Porter's Five Forces diagram with editable force descriptions",
+    "Use to analyze competitive dynamics using Porter's framework",
+    ["porter", "five forces", "cinque forze", "strategy", "strategia", "competition"]),
+  c("MnRiskMatrix", "mn-risk-matrix", "strategy",
+    "Risk matrix heatmap plotting likelihood vs. impact with editable risks",
+    "Use to assess and prioritize risks by probability and severity",
+    ["risk", "matrix", "rischio", "matrice", "likelihood", "probabilità", "impact", "impatto"]),
+  c("MnStrategyCanvas", "mn-strategy-canvas", "strategy",
+    "Strategy canvas line chart comparing value curves across competitors",
+    "Use to contrast strategic profiles using Blue Ocean value curves",
+    ["strategy", "canvas", "strategia", "tela", "value curve", "curva di valore", "comparison"]),
+  c("MnSwot", "mn-swot", "strategy",
+    "SWOT analysis grid with four editable quadrants",
+    "Use to conduct a structured SWOT analysis with collaborative editing",
+    ["swot", "strengths", "weaknesses", "opportunities", "threats", "forze", "debolezze"]),
+
+  // ── theme (6) ──
+  c("MnA11yFab", "mn-a11y-fab", "theme",
+    "Floating action button toggling accessibility overlays and settings",
+    "Use to give users quick access to accessibility tools from any page",
+    ["a11y", "fab", "accessibility", "accessibilità", "floating", "button", "pulsante"]),
+  c("MnA11y", "mn-a11y", "theme",
+    "Accessibility settings panel with contrast, font-size, and motion controls",
+    "Use to let users customize accessibility preferences app-wide",
+    ["a11y", "accessibility", "accessibilità", "contrast", "contrasto", "font", "motion"]),
+  c("MnDropdownMenu", "mn-dropdown-menu", "theme",
+    "Dropdown menu with nested sub-menus, icons, and keyboard navigation",
+    "Use for context menus or action menus triggered by a button",
+    ["dropdown", "menu", "tendina", "menù", "context", "contesto", "action", "azione"]),
+  { name: "MnManettino", slug: "mn-manettino", category: "theme",
+    description: "Ferrari-inspired rotary mode selector with haptic-style detent positions",
+    whenToUse: "Use as a themed mode or level selector with rotary dial interaction",
+    filePath: "theme/mn-ferrari-control.tsx", propsInterface: "MnManettinoProps",
+    keywords: ["manettino", "rotary", "ferrari", "selector", "selettore", "dial", "ghiera"] },
+  c("MnThemeRotary", "mn-theme-rotary", "theme",
+    "Rotary dial for selecting the active color theme with preview swatches",
+    "Use to let users browse and switch themes via a physical-dial metaphor",
+    ["theme", "rotary", "tema", "rotante", "color", "colore", "palette", "tavolozza"]),
+  c("MnThemeToggle", "mn-theme-toggle", "theme",
+    "Light/dark theme toggle switch with system-preference detection",
+    "Use to switch between light, dark, and system color schemes",
+    ["theme", "toggle", "tema", "interruttore", "dark", "light", "scuro", "chiaro", "mode"]),
+];
